+++ test/main/Test/LocalTime/TimeOfDay.hs
@@ -3,7 +3,6 @@ module Test.LocalTime.TimeOfDay
     ) where
 
 import Data.Time.LocalTime.Compat
-
 import Test.Arbitrary ()
 import Test.Tasty
 import Test.Tasty.QuickCheck hiding (reason)